+++ interfaces/lbInterfaces-sub-transfer.h
@@ -392,15 +392,15 @@ public:
 	virtual lbErrCodes LB_STDCALL recvInteger(int& i) = 0;
 	virtual lbErrCodes LB_STDCALL sendInteger(int i) = 0;
 	
-	virtual lbErrCodes LB_STDCALL send(void* buf, int len) = 0;
+	virtual lbErrCodes LB_STDCALL send(void* buf, short len) = 0;
 
 	/**
 	 * Buffer must be allocated.
 	 */	
-	virtual lbErrCodes LB_STDCALL recv(void* buf, int & len) = 0;
+	virtual lbErrCodes LB_STDCALL recv(void* buf, short & len) = 0;
 		
         virtual lbErrCodes LB_STDCALL recv_charbuf(char *buf) = 0;
-        virtual lbErrCodes LB_STDCALL send_charbuf(char *buf, int len) = 0;
+        virtual lbErrCodes LB_STDCALL send_charbuf(char *buf, short len) = 0;
 
 	/**
 	 * Send and recieve a data buffer and automatically split off to
